requirements.txt
absl-py==0.9.0
astor==0.8.1
bert-tensorflow==1.0.1
certifi==2020.4.5.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.0
opt-einsum==3.2.1
pandas==1.0.5
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
